requirements.txt
asgiref==3.2.7
Django==3.0.5
django-cors==0.1
django-cors-headers==3.5.0
djangorestframework==3.11.0
pytz==2020.1
sqlparse==0.3.1
virtualenv==16.7.8
